add comment submit

+++ src/main/java/life/majiang/community/dto/QuestionDTO.java
@@ -10,8 +10,8 @@ public class QuestionDTO {
     private String description;
     private String tag;
     private Long gmtCreate;
-    private  Long gmtModified;
-    private Integer creator;
+    private Long gmtModified;
+    private Long creator;
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
